only delete resources on Detach if created in Attach
We want to delete resources that were only created to facilitate an Attach()
request. The way this was determined was by checking if the given resource is
diskless or not. This presents a problem, as there are a number of ways
diskless resources get created:
* Linstor creates a TieBreaker resource
* The user specified that a Diskless resource should be placed on that node
* Attach() was called for a node were the resource was not present

From the point of view of Detach(), all these above cases look the same, and
so the resources were deleted even if they did not originate from the Attach.

This commit adds a new property to resources generated by Attach. This way
Detach can later check for the presence of this property, and only delete
the resource if:
* The resource has the "Attach" property
* The resource is still diskless (there could be cases in which LINSTOR
  migrated some replicas from a evicted node) Since we don't want to mess
  with the replication factor, diskfull resources are off-limits.

+++ pkg/client/linstor.go
@@ -414,35 +414,36 @@ func (s *Linstor) Attach(ctx context.Context, vol *volume.Info, node string) err
 	if err != nil {
 		return err
 	}
+
+	rc.Resource.Props[linstor.PropertyCreatedFor] = linstor.CreatedForTemporaryDisklessAttach
 	return s.client.Resources.Create(ctx, rc)
 }
 
 // Detach removes a volume from the node.
 func (s *Linstor) Detach(ctx context.Context, vol *volume.Info, node string) error {
+	log := s.log.WithFields(logrus.Fields{
+		"volume":     fmt.Sprintf("%+v", vol),
+		"targetNode": node,
+	})
+
 	res, err := s.client.Resources.Get(ctx, vol.ID, node)
-	if err == lapi.NotFoundError {
-		s.log.WithFields(logrus.Fields{
-			"volume":     fmt.Sprintf("%+v", vol),
-			"targetNode": node,
-		}).Info("volume does not even exist")
-		return nil
-	} else if err != nil {
-		return err
+	if err != nil {
+		// if the resource could not be found there is nothing to detach
+		return nil404(err)
 	}
 
-	s.log.WithFields(logrus.Fields{
-		"resource":   fmt.Sprintf("%+v", res),
-		"targetNode": node,
-	}).Info("detaching volume")
+	createdFor, ok := res.Props[linstor.PropertyCreatedFor]
+	if !ok || createdFor != linstor.CreatedForTemporaryDisklessAttach {
+		log.Info("resource not temporary (not created by Attach) not deleting")
+		return nil
+	}
 
 	if util.DeployedDiskfully(res) {
-		s.log.WithFields(logrus.Fields{
-			"resource":   fmt.Sprintf("%+v", res),
-			"targetNode": node,
-		}).Info("volume is diskfull on node, refusing to detach")
+		log.Info("temporary resource created by Attach is now diskfull, not deleting")
 		return nil
 	}
-
+	
+	log.Info("removing temporary resource")
 	return s.client.Resources.Delete(ctx, vol.ID, node)
 }
